+++ frontend/src/app/components/bible/verse-range-picker/verse-range-picker.component.ts
@@ -11,8 +11,8 @@ import {
 } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { UserService } from '../../../services/user.service';
-import { BibleService } from '../../../services/bible.service';
+import { UserService } from '@services/api/user.service';
+import { BibleService } from '@services/api/bible.service';
 
 export interface VerseSelection {
   mode: 'single' | 'range' | 'chapter';
