+++ frontend/src/app/components/memorization/memorization-modal/components/navigation-controls/navigation-controls.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input, Output, EventEmitter, OnInit, OnDestroy } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { trigger, transition, style, animate } from '@angular/animations';
-import { RecordingService, RecordingState } from '../../../../../services/recording.service';
-import { NotificationService } from '../../../../../services/notification.service';
+import { RecordingService, RecordingState } from '@services/utils/recording.service';
+import { NotificationService } from '@services/utils/notification.service';
 import { Subject, takeUntil } from 'rxjs';
 
 @Component({
