+++ frontend/src/app/components/memorization/memorization-modal/memorization-modal.component.ts
@@ -9,10 +9,10 @@ import {
 } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Router } from '@angular/router';
-import { BibleService } from '../../../services/bible.service';
-import { UserService } from '../../../services/user.service';
-import { RecordingService } from '../../../services/recording.service';
-import { BibleBook } from '../../../models/bible';
+import { BibleService } from '@services/api/bible.service';
+import { UserService } from '@services/api/user.service';
+import { RecordingService } from '@services/utils/recording.service';
+import { BibleBook } from '@models/bible';
 import { Subject, takeUntil } from 'rxjs';
 import { trigger, style, transition, animate, keyframes, state } from '@angular/animations';
