+++ frontend/src/app/features/feature-request/components/request-modal/request-modal.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input, Output, EventEmitter, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { CreateFeatureRequest, RequestType } from '../../../../models/feature-request.model';
-import { FeatureRequestService } from '../../../../services/feature-request.service';
+import { CreateFeatureRequest, RequestType } from '@models/feature-request.model';
+import { FeatureRequestService } from '@services/api/feature-request.service';
 
 @Component({
   selector: 'app-request-modal',
