+++ frontend/src/app/features/flow-memorization/services/flow-memorization.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Subject, debounceTime } from 'rxjs';
-import { BibleService } from '../../../services/bible.service';
+import { BibleService } from '@services/api/bible.service';
 import { FlowVerse } from '../models/flow.models';
 
 @Injectable()
