+++ frontend/src/app/features/profile/profile.component.ts
@@ -3,10 +3,10 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { RouterModule, Router } from '@angular/router';
-import { ModalService } from '../../services/modal.service';
-import { User } from '../../models/user';
-import { UserService } from '../../services/user.service';
-import { BibleService } from '../../services/bible.service';
+import { ModalService } from '@services/utils/modal.service';
+import { User } from '@models/user';
+import { UserService } from '@services/api/user.service';
+import { BibleService } from '@services/api/bible.service';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '../../../environments/environment';
 import { Subject, takeUntil } from 'rxjs';
