+++ frontend/src/app/features/scripture-atlas/services/map.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HistoricalBuildingsService } from './historical-buildings.service';
-import { ConfigService } from '../../../services/config.service';
+import { ConfigService } from '@services/config/config.service';
 
 @Injectable({
   providedIn: 'root'
