+++ frontend/src/app/layouts/components/navigation/navigation.component.ts
@@ -2,8 +2,8 @@
 import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Router, RouterModule } from '@angular/router';
-import { UserService } from '../../../services/user.service';
-import { User } from '../../../models/user';
+import { UserService } from '@services/api/user.service';
+import { User } from '@models/user';
 
 @Component({
   selector: 'app-navigation',
