+++ frontend/src/app/services/api/deck.service.ts
@@ -1,8 +1,8 @@
-// frontend/src/app/services/deck.service.ts
+// src/app/services/api/deck.service.ts
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable, tap, catchError, of } from 'rxjs';
-import { environment } from '../../environments/environment';
+import { environment } from '../../../environments/environment';
 import {
   Deck,
   Card,
@@ -11,7 +11,7 @@ import {
   UpdateDeckRequest,
   StudySessionResult,
   ImportDeckRequest,
-} from '../state/decks/models/deck.model';
+} from '@state/decks/models/deck.model';
 
 export interface DeckCreate {
   name: string;
