+++ frontend/src/app/services/auth/auth.service.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({ providedIn: 'root' })
+export class AuthService {
+  // Placeholder for authentication logic
+  constructor() {}
+}
