+++ frontend/src/app/state/decks/effects/deck.effects.ts
@@ -3,8 +3,8 @@ import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { Store } from '@ngrx/store';
 import { map, mergeMap, tap, filter, debounceTime } from 'rxjs/operators';
 
-import { DeckService } from '../../../services/deck.service';
-import { NotificationService } from '../../../services/notification.service';
+import { DeckService } from '@services/api/deck.service';
+import { NotificationService } from '@services/utils/notification.service';
 import { DeckActions, CardActions } from '../actions/deck.actions';
 import { Deck } from '../models/deck.model';
 import { BaseEffects } from '../../core/effects/base.effect';
