+++ frontend/src/app/state/practice-session/effects/practice-session.effects.ts
@@ -25,9 +25,9 @@ import {
 } from '../selectors/practice-session.selectors';
 import { BaseEffects } from '../../core/effects/base.effect';
 import { ResponseQuality } from '../models/practice-session.model';
-import { PracticeService } from '@services/practice.service';
-import { AudioService } from '@services/audio.service';
-import { NotificationService } from '@services/notification.service';
+import { PracticeService } from '@services/api/practice.service';
+import { AudioService } from '@services/utils/audio.service';
+import { NotificationService } from '@services/utils/notification.service';
 
 @Injectable()
 export class PracticeSessionEffects extends BaseEffects {
